#big_major_update #modules
This new version permit to use any module without using modules.json file. The program will be able to detect if a module is initialized, if need to be reconfigured etc...
A documentation will be writed soon to explain the strengths and weakness of the new mechanism.
Don't hesite to contact me by GitHub or by discord !